[AIR-1994] n10n: relocate types from impl.go to types.go
